+++ cou.h
@@ -9,7 +9,7 @@
 #include "util.h"
 
 typedef struct {
-    int db_size;
+    size_t db_size;
     char *db_cou_primary;
     char *db_cou_shandow;
     unsigned char *db_cou_curBA;
@@ -20,9 +20,9 @@ typedef struct {
 
 } db_cou_infomation;
 
-int db_cou_init(void *cou_info, int db_size);
-void *cou_read(int index);
-int cou_write(int index, void *value);
+int db_cou_init(void *cou_info, size_t db_size);
+void *cou_read(size_t index);
+int cou_write(size_t index, void *value);
 void ckp_cou(int ckp_id, void *cou_info);
 void db_cou_destroy(void *cou_info);
