+++ naive.h
@@ -10,16 +10,16 @@
 
 
 typedef struct {
-    int db_size;
+    size_t db_size;
     char *db_naive_AS;
     char *db_naive_AS_shandow;
 
     unsigned char db_naive_lock;
 } db_naive_infomation;
 
-int db_naive_init(void *db_naive_info, int db_size);
-void* naive_read(int index);
-int naive_write(int index, void *value);
+int db_naive_init(void *db_naive_info, size_t db_size);
+void* naive_read(size_t index);
+int naive_write(size_t index, void *value);
 void ckp_naive(int ckp_id, void *db_naive_info);
 void db_naive_destroy(void *db_naive_info);
