+++ pingpong.h
@@ -9,7 +9,7 @@
 #include<pthread.h>
 
 typedef struct {
-    int db_size;
+    size_t db_size;
     char *db_pp_as;
     char *db_pp_as_odd;
     char *db_pp_as_even;
@@ -19,9 +19,9 @@ typedef struct {
     unsigned char db_pp_lock;
     int current;
 } db_pingpong_infomation;
-int db_pingpong_init(void *pp_info, int db_size);
-void* pingpong_read(int index);
-int pingpong_write(int index, void *value);
+int db_pingpong_init(void *pp_info, size_t db_size);
+void* pingpong_read(size_t index);
+int pingpong_write(size_t index, void *value);
 void db_pingpong_ckp(int ckp_order, void *pp_info);
 void db_pingpong_destroy(void *pp_info);
