+++ zigzag.h
@@ -8,7 +8,7 @@
 #include<pthread.h>
 
 typedef struct {
-    int db_size;
+    size_t db_size;
     char *db_zigzag_as0;
     char *db_zigzag_as1;
     unsigned char *db_zigzag_mr;
@@ -16,9 +16,9 @@ typedef struct {
     unsigned char db_zigzag_lock;
 
 } db_zigzag_infomation;
-int db_zigzag_init(void *cou_info, int db_size);
-void* zigzag_read(int index);
-int zigzag_write(int index, void *value);
+int db_zigzag_init(void *cou_info, size_t db_size);
+void* zigzag_read(size_t index);
+int zigzag_write(size_t index, void *value);
 void db_zigzag_ckp(int ckp_order, void *cou_info);
 void db_zigzag_destroy(void *cou_info);
